refactor: redux features, fix: download for 1 file
