+++ package.json
@@ -8,7 +8,7 @@
         "build": "tsc && vite build",
         "preview": "vite preview",
         "test": "vitest run --coverage",
-        "vitest": "vitest run",
+        "vitest": "vitest",
         "format": "prettier --write .",
         "format:fix": "npx prettier . --write",
         "lint": "eslint .",
