+++ src/utils/test-utils.tsx
@@ -5,6 +5,7 @@ import type { PropsWithChildren, ReactElement } from 'react'
 import { Provider } from 'react-redux'
 import type { AppStore, RootState } from '../app/store'
 import { makeStore } from '../app/store'
+import { BrowserRouter } from 'react-router-dom'
 
 /**
  * This type extends the default options for
@@ -53,7 +54,9 @@ export const renderWithProviders = (
     } = extendedRenderOptions
 
     const Wrapper = ({ children }: PropsWithChildren) => (
-        <Provider store={store}>{children}</Provider>
+        <BrowserRouter>
+            <Provider store={store}>{children}</Provider>
+        </BrowserRouter>
     )
 
     // Return an object with the store and all of RTL's query functions
